用htmx移除comments

+++ comments/views.py
@@ -4,6 +4,7 @@ from articles.models import Article
 from django.contrib import messages
 from comments.models import Comment
 from django.views.decorators.http import require_POST
+from django.http import HttpResponse
 
 def create(request, id):
     if request.method == 'POST':
@@ -18,9 +19,11 @@ def create(request, id):
         return redirect("articles:detail", article.id)
 
 
-@require_POST
 def delete(request, id):
-    comment = get_object_or_404(Comment, pk=id)
-    comment.delete()
-    messages.warning(request, "刪除留言成功")
-    return redirect("articles:detail", comment.article_id)
+    if request.method == "DELETE":
+        comment = get_object_or_404(Comment, pk=id)
+        comment.delete()
+        # messages.warning(request, "刪除留言成功")
+
+        return HttpResponse("")
+        # return redirect("articles:detail", comment.article_id)
